ADD: spec updates

+++ backends/foundation_jsonschema/src/types.rs
@@ -285,7 +285,7 @@ mod tests {
 
     #[test]
     fn test_json_type_of_number() {
-        let n: Value = serde_json::Number::from_f64(std::f32::consts::PI)
+        let n: Value = serde_json::Number::from_f64(std::f64::consts::PI)
             .unwrap()
             .into();
         assert_eq!(JsonType::of(&n), JsonType::Number);
